CH: Moving non-page components to new folder
