+++ frontend/src/pages/CreateGroup.tsx
@@ -13,7 +13,7 @@ import { auth, db, DB_GROUP_COLLECT, DB_KEY_SLUG_OPTS } from "../firebase";
 import { useNavigate } from "react-router-dom";
 import { get, query, ref, set } from "firebase/database";
 import slugify from "slugify";
-import Header from "./Header";
+import Header from "../components/Header";
 import DropZone, { storage } from "../storage";
 import { uploadBytes } from "firebase/storage";
 import { ref as storRef } from "firebase/storage";
