+++ frontend/src/pages/GroupPage.tsx
@@ -16,7 +16,7 @@ import { ref } from "firebase/database";
 import { Group } from "./CreateGroup";
 import { Val } from "react-firebase-hooks/database/dist/database/types";
 import { useList, useObjectVal } from "react-firebase-hooks/database";
-import Header from "./Header";
+import Header from "../components/Header";
 import { storage } from "../storage";
 import { useDownloadURL } from "react-firebase-hooks/storage";
 import { ref as storageRef } from "firebase/storage";
