+++ frontend/src/pages/JoinGroup.tsx
@@ -6,7 +6,7 @@ import { Group } from "./CreateGroup";
 import { ref, set } from "firebase/database";
 import { auth, db } from "../firebase";
 import { useAuthState } from "react-firebase-hooks/auth";
-import Header from "./Header";
+import Header from "../components/Header";
 
 const FoundGroup = ({ group, userId }: { group: Group; userId: string }) => {
   const navigate = useNavigate();
